had to update the way I was presenting pageHeader but also slugs for the projects worK
