+++ app/page.tsx
@@ -1,6 +1,6 @@
 import About from '@/src/components/About/About';
 import FeaturedProjects from '@/src/components/Projects/FeaturedProjects';
-import PageHeader from '@/src/components/PageHeader/PageHeader';
+import HomePageHeader from '@/src/components/Home/HomePagerHeader';
 import Quotes from '@/src/components/Quotes/Quotes';
 import WebsiteStatusCard from '@/src/components/WebsiteStatus/WebsiteStatusCard';
 import SubmitFeedbackCard from '@/src/components/SubmitFeedback/SubmitFeedbackCard';
@@ -9,7 +9,7 @@ import SpotifyStatusCard from '@/src/components/Spotify/SpotifyStatusCard';
 export default function Home() {
   return (
     <>
-      <PageHeader />
+      <HomePageHeader />
 
       <div className="grid grid-cols-1 md:grid-cols-3 gap-y-4 md:gap-4 mb-20">
         {/* Main Content */}
