+++ app/projects/[project]/page.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+export default function ProjectPage({
+  params,
+}: {
+  params: { project: string };
+}) {
+  return (
+    <div>
+      <p>
+        Welcome to the <strong>{params.project}</strong> project!
+      </p>
+    </div>
+  );
+}
