+++ src/components/PageHeader/PageHeader.tsx
@@ -1,37 +1,21 @@
 'use client';
 import React from 'react';
 
-const PageHeader = () => {
-  const clientTimeZone = Intl.DateTimeFormat().resolvedOptions().timeZone;
-  const getHeader = () => {
-    const hours = new Date().toLocaleString('en-US', {
-      timeZone: clientTimeZone,
-      hour: '2-digit',
-      hour12: false,
-    });
-    const numericHours = parseInt(hours, 10);
-    if (numericHours < 12) return 'Good Morning!';
-    else if (numericHours < 18) return 'Good Afternoon!';
-    return 'Good Evening!';
-  };
+interface PageHeaderProps {
+  header: string | (() => string);
+  subHeader: string | (() => string);
+}
 
-  const getSubHeader = () => {
-    const date = new Date();
-    const formattedDate = new Intl.DateTimeFormat('en-US', {
-      weekday: 'long',
-      month: 'long',
-      day: 'numeric',
-      year: 'numeric',
-      timeZone: clientTimeZone,
-    }).format(date);
-
-    return `It's ${formattedDate}`;
+const PageHeader: React.FC<PageHeaderProps> = ({ header, subHeader }) => {
+  // Function to resolve the prop value if it's a function or return the string directly
+  const resolveValue = (value: string | (() => string)) => {
+    return typeof value === 'function' ? value() : value;
   };
 
   return (
     <div className="page-header">
-      <h1 className="text-4xl font-bold">{getHeader()}</h1>
-      <h3 className="text-2xl secondary-font">{getSubHeader()}</h3>
+      <h1 className="text-4xl font-bold">{resolveValue(header)}</h1>
+      <h3 className="text-2xl secondary-font">{resolveValue(subHeader)}</h3>
     </div>
   );
 };
